pyproject.toml
[tool.poetry]
name = "ot-harjoitustyo"
version = "0.1.0"
description = ""
authors = ["Jaakko Airikkala <jaakkel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
invoke = "^2.2.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
coverage = "^7.6.7"
pylint = "^3.3.1"
autopep8 = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
